src: add node_process.cc
Backport of https://github.com/nodejs/node/pull/21105
